Remove OS/2 - Windows type usage from portable code This commit finishes all the CHAR and related (PSZ, etc.) types.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1094 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/mark/os2-types/trunk/kernel/instructions/CallInstruction.cpp
@@ -59,8 +59,8 @@ RexxInstructionCall::RexxInstructionCall(
     RexxString *condition,             /* CALL ON/OFF condition             */
     size_t      argCount,              /* count of arguments                */
     RexxQueue  *argList,               /* call arguments                    */
-    uint8_t flags,                     /* CALL flags                        */
-    uint8_t builtin_index)             /* builtin routine index             */
+    size_t      flags,                 /* CALL flags                        */
+    size_t      builtin_index)         /* builtin routine index             */
 /******************************************************************************/
 /* Function:  Complete CALL instruction object                                */
 /******************************************************************************/
