+++ sandbox/mark/os2-types/trunk/kernel/instructions/NumericInstruction.hpp
@@ -51,7 +51,7 @@
 class RexxInstructionNumeric : public RexxInstructionExpression {
  public:
   inline void *operator new(size_t size, void *ptr) {return ptr;};
-  RexxInstructionNumeric(RexxObject *, USHORT, uint8_t);
+  RexxInstructionNumeric(RexxObject *, USHORT, size_t);
   inline RexxInstructionNumeric(RESTORETYPE restoreType) { ; };
   void execute(RexxActivation *, RexxExpressionStack *);
 };
