+++ sandbox/mark/os2-types/trunk/kernel/instructions/TraceInstruction.cpp
@@ -51,7 +51,7 @@
 RexxInstructionTrace::RexxInstructionTrace(
     RexxObject *_expression,           /* TRACE VALUE expression            */
     USHORT      trace,                 /* trace setting                     */
-    uint8_t     debug_flags,           /* new debug setting                 */
+    size_t      debug_flags,           /* new debug setting                 */
     LONG        debug_skip )           /* debug clauses to skip             */
 /******************************************************************************/
 /* Function:  Initialize a trace instruction                                  */
